+++ src/main/java/com/backend/fluxnewsapi/services/UsersRepository.java
@@ -1,10 +1,11 @@
 package com.backend.fluxnewsapi.services;
 
-import com.backend.fluxnewsapi.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import com.backend.fluxnewsapi.models.User;
 
-public interface UsersRepository extends JpaRepository<User,Integer> {
+
+public interface UsersRepository extends JpaRepository<User,Long> {
     //List<ArticleUser> findByArticleUsers(int id);
     User findByEmail(String email);
     User findByUsername(String username);
